backend/requirements.txt
Flask==2.2.3
pymongo==4.3.3
elasticsearch==8.6.2
python-dotenv==1.0.0
Flask-Cors==3.0.10
bson==0.5.10